Moved navigation to components in Navigation.js updated package.JSON and package-lock.json removed index.js
