+++ App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import AppContainer from './Index';
+import AppContainer from './app/components/CustomNavigation/Navigation';
 
 export default class App extends React.Component {
   render() {
