+++ app/components/CustomNavigation/Navigation.js
@@ -1,8 +1,8 @@
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 import { createStackNavigator } from 'react-navigation-stack';
-import HomeScreen from './app/view/HomeScreen/HomeScreen';
-import MainScreen from './app/view/MainScreen/MainScreen';
-import LoadingScreen from './app/view/LoadingScreen/LoadingScreen';
+import HomeScreen from '../../view/HomeScreen/HomeScreen';
+import MainScreen from '../../view/MainScreen/MainScreen';
+import LoadingScreen from '../../view/LoadingScreen/LoadingScreen';
 
 const RootStack = createStackNavigator(
     {
