+++ package.json
@@ -19,7 +19,9 @@
     "react-native-status-bar-height": "^2.4.0",
     "react-native-web": "^0.11.7",
     "react-navigation": "^4.0.10",
-    "react-navigation-stack": "^1.10.3"
+    "react-navigation-stack": "^1.10.3",
+    "react-redux": "^7.1.1",
+    "redux": "^4.0.4"
   },
   "devDependencies": {
     "babel-preset-expo": "^7.1.0"
